feat: Display photos and allow deleting
Photos will now display in the collection view and can be deleted.
The photos as downloaded when a new pin is dropped, however before anything is
visible in the Collection View all the data needs to download first. Need to try and get the data to download in the background